Refactoring of package structure
Three new packages introduced: core, dd, and matrix.
Examples test directory can now be found in the it source root.
