+++ src/main/scala/com/phasmid/majabigwaduce/core/FP.scala
@@ -1,0 +1,5 @@
+package com.phasmid.majabigwaduce.core
+
+object FP {
+
+}
