+++ src/main/scala/com/phasmid/majabigwaduce/core/LazyMonad.scala
@@ -1,0 +1,5 @@
+package com.phasmid.majabigwaduce.core
+
+trait LazyMonad {
+
+}
