+++ src/main/scala/com/phasmid/majabigwaduce/core/Mapper.scala
